+++ CloudShell.ResourceDefinitions/ResourceResolver.cs
@@ -408,6 +408,8 @@ public static class ResourceDefinitionDiagnosticCodes
     public const string ResourceProjectionProviderMissing = "resourceDefinition.projectionProviderMissing";
     public const string ResourceChangeApplyProviderMissing = "resourceDefinition.changeApplyProviderMissing";
     public const string ResourceGraphVersionConflict = "resourceDefinition.graphVersionConflict";
+    public const string ResourceGraphResourceMissing = "resourceDefinition.graphResourceMissing";
+    public const string ResourceDependencyCycle = "resourceDefinition.dependencyCycle";
 }
 
 public static class ResourceDefinitionClassIds